Water absorbs into ground before it pools; washing; 32-bit save words
Ground now drinks water before water persists on top of it, authored per
substrate rather than per liquid.

- "absorb": {"capacity": N} on a SOLID/POWDER says how deep a stain that
  material will hold (grass 12, dirt 10, ash 14, sand 6; absent = 0 = never
  absorbs, which is all stone). The liquid's `amount` stays the per-contact
  STEP and the substrate's capacity is the CEILING — two genuinely different
  axes, and authoring the ceiling per material PAIR would be the N x M
  explosion tags exist to avoid. Effective depth is min(amount, capacity).

- Absorbing SPENDS the liquid: one eighth of the source cell's fullness per
  contact, in the same units stepLiquid speaks, and the cell dies when it gives
  its last. Without that debit the puddle stains the ground and then sits there
  full forever, which is not absorption. The stain climbs one level per eighth
  spent, so ground visibly darkens as it drinks and the depth reached is paid
  for in real mass.

- "washes": true on water rinses FOREIGN stains out (steps the amount down one
  level per contact) instead of overwriting them. Water over blood-soaked
  ground would otherwise relabel the blood as "wet" at full strength: the
  colour would change but the mess would never come out.

Both fit in stainPack's spare bits (27..30 capacity, 31 washes), so the 64-byte
Material did not grow. Rule 2 holds because every step is monotone toward a
bounded fixed point — stain rises only to the ceiling, a washed stain falls only
to 0, absorbed fullness falls only to 0 — so a saturated puddle on saturated
ground sleeps. Saturation is deliberately terminal: a cell that could both wet
and dry would never reach a fixed point.

Two pre-existing bugs surfaced because a second stain type now exists:

- The save format dropped every stain. RleEncodeChunk persisted only the low
  16 bits, but stain lives at bits 24..30 and IS folded into the determinism
  hash, so a saved-and-reloaded world hashed differently from the one saved.
  Invisible while blood was the only stainer (nothing was stained at save
  time); a hard save/load failure once worldgen ponds wet their banks. The
  persisted word is now 32-bit masked by kPersistMask (everything but the stamp
  byte) and the region magic went 'SVR1' -> 'SVR2' so old files are rejected
  rather than misread at the wrong stride.

- The blood-stain selftest swept the whole world asserting every stain was
  blood, so 24052 legitimate "wet" stains read as malformed packing. It now
  asserts well-formedness; the blood-specific claim stays scoped to its floor.

selftest: determinism PASS (87a9cd11, stable across runs), sleep PASS (0/32768
active), blood stain PASS (0 malformed), save/load PASS (7d6d6d2f restored
exactly), region store / streaming / perf PASS. pond freeze and micro body
render still FAIL identically on clean c335d8a — pre-existing, not from this
change.

+++ src/sim/chunkstore.h
@@ -22,15 +22,15 @@ class ChunkStore {
  public:
   struct Entry {
     IVec3 wc;
-    std::vector<uint16_t> rle;
+    std::vector<uint32_t> rle;
   };
   static constexpr int kRegionShift = 4;  // 16 chunks (256 voxels) per axis
   static constexpr size_t kMaxRamRegions = 64;
 
-  void Put(IVec3 wc, std::vector<uint16_t> rle);
+  void Put(IVec3 wc, std::vector<uint32_t> rle);
   // Pointer is valid only until the next Put/Get/Clear: either may LRU-spill
   // the region that owns it. Use immediately.
-  const std::vector<uint16_t>* Get(IVec3 wc);
+  const std::vector<uint32_t>* Get(IVec3 wc);
 
   // Forget everything in RAM and detach from any bound directory. Files are
   // left on disk untouched (a regen must not destroy the last explicit save;
